+++ lib/torngen/spec/parameter_content.ex
@@ -2,7 +2,8 @@ defmodule Torngen.Spec.Parameter.Content do
   @moduledoc """
   Represents content parameters for OpenAPI requests.
 
-  This struct is intended to extend the base `Torngen.Spec.Parameter` struct when handling parameters for OpenAPI requests.
+  This struct is intended to extend the base `Torngen.Spec.Parameter` struct when handling
+  parameters for OpenAPI requests.
 
   ## External Resources
   - [Swagger Specification](https://swagger.io/specification#fixed-fields-for-use-with-content)
@@ -15,7 +16,8 @@ defmodule Torngen.Spec.Parameter.Content do
 
   ## Fields
 
-    * `content` - A map containing the representations for the parameter. The key is the media type and the value describes it.
+    * `content` - A map containing the representations for the parameter. The key is the media type
+       and the value describes it.
   """
   @type t :: %__MODULE__{
           content: map()